refactor: BaseActivity를 사용하도록 변경
