+++ android/app/src/main/java/io/coursepick/coursepick/presentation/preference/PreferencesActivity.kt
@@ -3,12 +3,12 @@ package io.coursepick.coursepick.presentation.preference
 import android.os.Bundle
 import android.view.View
 import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import io.coursepick.coursepick.R
+import io.coursepick.coursepick.presentation.base.BaseActivity
 
-class PreferencesActivity : AppCompatActivity() {
+class PreferencesActivity : BaseActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
